manager process

+++ php/processloanreq.php
@@ -5,7 +5,7 @@ $postdata = file_get_contents('php://input');
 if(isset($postdata) && !empty($postdata))
 {
 	$request = json_decode($postdata,true);
-	if (trim($request['reqno']) == '' || (float)$request['accountno'] < 0) {
+	if (trim($request['reqno']) == '' || (int)$request['accountno'] < 0) {
 		return http_response_code(400);
 	}
 	$reqno = mysqli_real_escape_string($mysqli, (int)$request['reqno']);
